Implemented suggestions
